Refactored some pathfinder classes and tests
