--- KioskApplication/src/KioskApplication/pathfinder/PathfindingException.java
@@ -1,8 +1,0 @@
-package KioskApplication.pathfinder;
-
-public class PathfindingException extends Exception{
-
-  public  PathfindingException(String message){
-      System.out.println(message);
-  }
-}
