+++ KioskApplication/src/KioskApplication/tests/pathfinderTests/TestPathfinderNode.java
@@ -1,6 +1,6 @@
 package KioskApplication.tests.pathfinderTests;
 
-public class pathfindingnodeTest {
+public class TestPathfinderNode {
